APM-2975 Split test_jwt_unattended_access into client_credentials and token_exchange
